Type fixes.

+++ cpp/dolfinx/graph/GraphBuilder.h
@@ -39,7 +39,7 @@ public:
   /// Build distributed dual graph (cell-cell connections) from minimal
   /// mesh data, and return (graph, ghost_vertices, [num local edges,
   /// num non-local edges])
-  static std::pair<std::vector<std::vector<std::size_t>>,
+  static std::pair<std::vector<std::vector<std::int64_t>>,
                    std::array<std::int32_t, 3>>
   compute_dual_graph(
       const MPI_Comm mpi_comm,
@@ -51,8 +51,8 @@ public:
   /// Compute local part of the dual graph, and return (local_graph,
   /// facet_cell_map, number of local edges in the graph (undirected)
   static std::tuple<
-      std::vector<std::vector<std::size_t>>,
-      std::vector<std::pair<std::vector<std::size_t>, std::int32_t>>,
+      std::vector<std::vector<std::int32_t>>,
+      std::vector<std::pair<std::vector<std::int32_t>, std::int32_t>>,
       std::int32_t>
   compute_local_dual_graph(
       const Eigen::Ref<const Eigen::Array<std::int64_t, Eigen::Dynamic,
